[Shi Yangyi] modify BigInteger to Long in project

+++ src/main/java/com/example/yangyistarter/controller/ParkingController.java
@@ -7,8 +7,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
-
 @RestController
 @AllArgsConstructor
 @RequestMapping("parking")
@@ -18,19 +16,19 @@ public class ParkingController {
 
     @PreAuthorize("hasRole('ROLE_MANAGER')")
     @PostMapping("/managers")
-    public BigInteger randomParking() {
+    public long randomParking() {
         return parkingService.randomParking();
     }
 
     @PreAuthorize("hasRole('ROLE_CLEVER_ASSISTANT')")
     @PostMapping("/clever/assistants")
-    public BigInteger cleverParking() {
+    public long cleverParking() {
         return parkingService.cleverParking();
     }
 
     @PreAuthorize("hasAnyRole('ROLE_STUPID_ASSISTANT', 'ROLE_USER')")
     @PostMapping("/users")
-    public BigInteger parking() {
+    public long parking() {
         return parkingService.parking();
     }
 }
