+++ src/main/java/com/example/yangyistarter/dto/UserDTO.java
@@ -4,16 +4,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
-
 import javax.validation.constraints.Pattern;
-import java.math.BigInteger;
 
 @Builder
 @Setter
 @Getter
 public class UserDTO {
 
-    private BigInteger id;
+    private Long id;
     @Length(max = 20, message = "用户名不能超过20个字符")
     @Length(min = 3, message = "用户名不能少于3个字符")
     @Pattern(regexp = "^[A-Za-z0-9\\-\\_]*$", message = "用户名只能包含字母、数字、-、_")
