+++ src/main/java/com/example/yangyistarter/repository/ParkingLotRepository.java
@@ -4,7 +4,6 @@ import com.example.yangyistarter.entity.ParkingLot;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 @Repository
-public interface ParkingLotRepository extends JpaRepository<ParkingLot, BigInteger> {
+public interface ParkingLotRepository extends JpaRepository<ParkingLot, Long> {
 }
