+++ src/main/java/com/example/yangyistarter/service/PickingService.java
@@ -5,8 +5,6 @@ import com.example.yangyistarter.repository.ParkingSpaceRepository;
 import com.example.yangyistarter.util.ResponseCode;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Service
@@ -15,7 +13,7 @@ public class PickingService {
 
     ParkingSpaceRepository parkingSpaceRepository;
 
-    public ResponseCode picking(BigInteger receiptId) {
+    public ResponseCode picking(long receiptId) {
         Optional<ParkingSpace> parkingSpace = parkingSpaceRepository.findById(receiptId);
         if (!parkingSpace.isPresent()) {
             return ResponseCode.RECEIPT_INVALID;
